src/microservices/usuario/requirements.txt
Django==1.11.10
djangorestframework==3.7.7
mongoengine==0.15.0
PyJWT==1.5.3
pymongo==3.6.1
pytz==2018.3
six==1.11.0
